refactored instance into amplitude client
