+++ test/tests.js
@@ -3,7 +3,8 @@
   require('./md5.js');
   require('./uuid.js');
   require('./cookie.js');
-  require('./amplitude.js');
+  require('./amplitude-client.js');
   require('./ua-parser.js');
   require('./identify.js');
   require('./cookiestorage.js');
+  require('./utm.js');
